+++ module/core/src/main/java/ru/protei/portal/core/service/WorkerPositionServiceImpl.java
@@ -2,6 +2,7 @@ package ru.protei.portal.core.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import ru.protei.portal.api.struct.Result;
+import ru.protei.portal.core.model.dao.WorkerEntryDAO;
 import ru.protei.portal.core.model.dao.WorkerPositionDAO;
 import ru.protei.portal.core.model.dict.En_ResultStatus;
 import ru.protei.portal.core.model.ent.AuthToken;
@@ -20,6 +21,9 @@ public class WorkerPositionServiceImpl implements WorkerPositionService{
     WorkerPositionDAO workerPositionDAO;
 
     @Autowired
+    WorkerEntryDAO workerEntryDAO;
+
+    @Autowired
     PolicyService policyService;
 
     @Override
@@ -29,7 +33,7 @@ public class WorkerPositionServiceImpl implements WorkerPositionService{
     }
 
     @Override
-    public Result<Long> createWorkerPositions(AuthToken token, WorkerPosition workerPosition) {
+    public Result<Long> createWorkerPosition(AuthToken token, WorkerPosition workerPosition) {
         if (!isValid(workerPosition)) {
             return error(En_ResultStatus.INCORRECT_PARAMS);
         }
@@ -44,7 +48,7 @@ public class WorkerPositionServiceImpl implements WorkerPositionService{
     }
 
     @Override
-    public Result<Long> updateWorkerPositions(AuthToken token, WorkerPosition workerPosition) {
+    public Result<Long> updateWorkerPosition(AuthToken token, WorkerPosition workerPosition) {
         if (!isValid(workerPosition)) {
             return error(En_ResultStatus.INCORRECT_PARAMS);
         }
@@ -57,7 +61,12 @@ public class WorkerPositionServiceImpl implements WorkerPositionService{
     }
 
     @Override
-    public Result<Long> removeWorkerPositions(AuthToken token, WorkerPosition workerPosition) {
+    public Result<Long> removeWorkerPosition(AuthToken token, WorkerPosition workerPosition) {
+
+        if(workerEntryDAO.checkExistsByPosId(workerPosition.getId())){
+            return error(En_ResultStatus.WORKER_WITH_THIS_POSITION_ALREADY_EXIST);
+        }
+
         boolean result = workerPositionDAO.removeByKey(workerPosition.getId());
 
         if ( !result )
